requirements.txt
flask
pandas
opencv-python-headless
matplotlib
psycopg2-binary

